Support zcrypt using OpenSSL DES
Support a hackish mechanism for detecting OpenSSL DES if des425 is not
present.  If you include des.h then they include macros to be
consistent with the des425 API.  However these macros rename
des_ecb_encrypto te DES_ecb_encrypt.  Searching for both the macros
and the functions was annoying in autoconf, so instead we search for
those internal function names.

Also, search  for /usr/include/openssl if no krb4 includes are found.