Build reply commands from scratch on outgoing zwrites
We currently reuse the original zwrite line which is problematic when
the user passed -m. This instead regenerates the zwrite line from the
parsed owl_zwrite.

This adds a new replyline attribute to store the outgoing reply command.

Alternatively, we can just reuse the zwriteline attribute. This would
result in a somewhat backwards-incompatible change in that zwriteline is
now "the zwriteline to reply here" rather than "the zwriteline that
created this message", however anyone using that attribute likely had
the same bug. Thoughts?