Remove default_{attrs,fgcolor,bgcolor} from owl_fmtext
They didn't quite behave when appending two owl_fmtexts together; what
does it mean to append a string whose default color is red to one whose
default color is blue. (And then you set the default to green
afterwards...)

This means owl_fmtext is merely a unicode string with formatting
characters.  Default attributes are instead a property of how you view
the string and are passed to owl_fmtext_curs_waddstr. Which
unfortunately means it has a lot of arguments. :-/