Apply default attributes after processing commands
Otherwise we reset the color after every clump of formatting characters,
whether it includes a color or not. (Though this never happens because
of how we generate fmtext.) Also, actually modify the current attr so
that default_attrs works with search.

(While I'm at it, simplify _owl_fmtext_append_fmtext a bit.)