Move log-writing onto a background thread.
Currently, Barnowl writes message logs on the same thread as the UI and other
I/O. This means that if writing log files is very slow (e.g. if logpath or
classlogpath point to somewhere in AFS and the fileserver that volume relies on
is very slow to respond), barnowl freezes up until the write completes. This
patch creates a background logging thread as part of initialization and has the
UI thread instead push log messages onto a GAsyncQueue for the logging thread to
pop off and write in the background.