Use a GMainLoop on the log-writing thread instead of an explict queue.
Right now, we write log files via a background thread that waits on messages to
write being placed in a GAsyncQueue for it. This patch removes the explicit
GAsyncQueue in favor of just running a main loop on that thread and posting
tasks to it (a la how other background threads can post tasks on the main
thread).